[MODIFIED] all float members modified to double type

+++ src/formation/java/tp/model/CD.java
@@ -8,11 +8,11 @@ import java.util.Date;
 
 public class CD extends ACollectible implements IDisplayable
 {
-    private float   mLength;
+    private double  mLength;
     private eCDType mCDType;
 
     public CD(){}
-    public CD(String pTitle, Editor pEditor, Date pPublishDate, float pLength, eCDType pCDType )
+    public CD(String pTitle, Editor pEditor, Date pPublishDate, double pLength, eCDType pCDType )
     {
         this.mPublishDate = pPublishDate ;
         this.mBorrowed    = false ;
@@ -28,7 +28,7 @@ public class CD extends ACollectible implements IDisplayable
     public void borrowDVD()      {this.mBorrowed = true ;}
     public Editor getEditor()    {return this.mEditor ;}
     public String getTitle()     {return this.mTitle ;}
-    public float getDVDLength()  {return this.mLength ;}
+    public double getDVDLength() {return this.mLength ;}
     public eCDType getCDType()   {return this.mCDType ;}
 
     @Override
